pyproject.toml
[tool.poetry]
name = "KarenBot"
version = "0.1.0"
description = ""
authors = ["syrinka <skycore@foxmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0-beta.4"
Flask = {extras = ["async"], version = "^2.1.3"}
nonebot-plugin-gocqhttp = "^0.5.6"
nonebot-plugin-help = "^0.3.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.7"

[tool.nonebot]
plugins = ["nonebot_plugin_gocqhttp", "nonebot_plugin_help"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
